+++ packages/verified-fetch/test/custom-dns-resolvers.spec.ts
@@ -32,10 +32,10 @@ describe('custom dns-resolvers', () => {
     await expect(response.text()).to.eventually.equal('hello world')
 
     expect(customDnsResolver.calledWith('_dnslink.some-non-cached-domain.com', {
-      isRawBlockRequest: false,
       types: [
         RecordType.TXT
-      ]
+      ],
+      logger: undefined
     })).to.be.true()
   })
 
@@ -63,10 +63,10 @@ describe('custom dns-resolvers', () => {
 
     expect(customDnsResolver.callCount).to.equal(1)
     expect(customDnsResolver.getCall(0).args).to.deep.equal(['_dnslink.some-non-cached-domain2.com', {
-      isRawBlockRequest: false,
       types: [
         RecordType.TXT
-      ]
+      ],
+      logger: undefined
     }])
   })
 })